Goal Complete - But barely
All portions of the goal are complete but I just met the bare minimum. While coding this, I found many areas for improvemnt. In an effort to keep the week timeline I stopped here. Would like to get some feedback on app and continue improving it. Needs to be improved by putting in failsafes for when user messes things up.